telegram-azure/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
anyio==4.6.2.post1
appnope==0.1.4
asttokens==3.0.0
attrs==24.2.0
azure-common==1.1.28
azure-core==1.32.0
azure-cosmos==4.9.0
azure-functions==1.21.3
azure-functions-durable==1.2.10
azure-identity==1.19.0
azure-mgmt-core==1.5.0
azure-mgmt-resource==23.2.0
azure-mgmt-storage==21.2.1
azure-mgmt-web==7.3.1
azure-storage-blob==12.24.1
azure-storage-file-share==12.20.1
bcrypt==4.2.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
comm==0.2.2
cryptography==44.0.0
debugpy==1.8.12
decorator==5.1.1
executing==2.2.0
frozenlist==1.5.0
furl==2.1.3
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
isodate==0.7.2
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
nest-asyncio==1.6.0
orderedmultidict==1.0.1
packaging==24.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
portalocker==2.10.1
prompt_toolkit==3.0.50
propcache==0.2.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
yarl==1.18.3
